feat: add async put operation for Node.js bindings
Implemented PutWorker class extending WaveDBAsyncWorker for async put operations:

- Created put_worker.h/cc with PutWorker class
- Added Put method to database.cc WaveDB class
- Created header files for utility functions (path.h, identifier.h, async_worker.h)
- Converted .cc includes to proper header includes
- Added placeholder files for future workers (get, del, batch, iterator)
- Added -fPIC flag to CMakeLists.txt for shared library support

The Put operation:
- Accepts key (string or array) and value (string or Buffer)
- Optional callback for callback-style API
- Returns Promise for promise-style API
- Properly cleans up path and value in destructor
- Uses database_put_sync for thread-safe async execution
- Supports error handling with error codes

Build system:
- Updated binding.gyp to include new worker files
- Fixed library path for linking with libwavedb.a
- Enabled -fPIC for position-independent code

+++ bindings/nodejs/src/path.cc
@@ -1,9 +1,7 @@
-#include <napi.h>
-#include <string>
-#include <vector>
-#include "../../../src/HBTrie/path.h"
+#include "path.h"
 #include "../../../src/HBTrie/identifier.h"
 #include "../../../src/Buffer/buffer.h"
+#include <cctype>
 
 // Split string by delimiter
 std::vector<std::string> SplitString(const std::string& str, char delimiter) {
@@ -102,7 +100,7 @@ path_t* PathFromJS(Napi::Env env, Napi::Value key, char delimiter) {
 std::string PathToJS(path_t* path, char delimiter) {
   std::string result;
 
-  for (size_t i = 0; i < path->identifiers.length; i++) {
+  for (size_t i = 0; i < static_cast<size_t>(path->identifiers.length); i++) {
     if (i > 0) {
       result += delimiter;
     }
@@ -110,7 +108,7 @@ std::string PathToJS(path_t* path, char delimiter) {
     identifier_t* id = path->identifiers.data[i];
 
     // Extract identifier bytes to string
-    for (size_t j = 0; j < id->chunks.length; j++) {
+    for (size_t j = 0; j < static_cast<size_t>(id->chunks.length); j++) {
       chunk_t* chunk = id->chunks.data[j];
       const uint8_t* data = static_cast<const uint8_t*>(chunk_data_const(chunk));
       size_t size = chunk->data->size;
@@ -144,12 +142,12 @@ std::string PathToJS(path_t* path, char delimiter) {
 Napi::Array PathToArrayJS(Napi::Env env, path_t* path, char delimiter) {
   Napi::Array arr = Napi::Array::New(env);
 
-  for (size_t i = 0; i < path->identifiers.length; i++) {
+  for (size_t i = 0; i < static_cast<size_t>(path->identifiers.length); i++) {
     identifier_t* id = path->identifiers.data[i];
 
     // Extract identifier as string
     std::string part;
-    for (size_t j = 0; j < id->chunks.length; j++) {
+    for (size_t j = 0; j < static_cast<size_t>(id->chunks.length); j++) {
       chunk_t* chunk = id->chunks.data[j];
       const uint8_t* data = static_cast<const uint8_t*>(chunk_data_const(chunk));
       size_t size = chunk->data->size;
